Revert "Move common dependencies to common-lib"
