+++ apps/checker/app/services/RuleProvisionerService.scala
@@ -3,14 +3,16 @@ package services
 import java.util.Date
 
 import akka.actor.Scheduler
-import com.gu.typerighter.model.{BaseRule, Category, LTRule, LTRuleXML, RegexRule, RuleResource}
-import com.gu.typerighter.rules.BucketRuleManager
 import matchers.RegexMatcher
 import play.api.Logging
 
 import scala.concurrent.ExecutionContext
 import scala.concurrent.duration._
+import model.{RegexRule, BaseRule, Category, RuleResource}
+import rules.BucketRuleManager
 import matchers.LanguageToolFactory
+import model.LTRule
+import model.LTRuleXML
 import utils.CloudWatchClient
 import utils.Metrics
