+++ apps/checker/app/services/SentenceHelpers.scala
@@ -1,6 +1,6 @@
 package services
 
-import com.gu.typerighter.model.TextRange
+import model.TextRange
 import opennlp.tools.sentdetect.{SentenceDetectorME, SentenceModel}
 import opennlp.tools.tokenize.SimpleTokenizer
 import opennlp.tools.util.Span
