+++ apps/checker/app/utils/Text.scala
@@ -1,4 +1,6 @@
-package com.gu.typerighter.model
+package utils
+
+import model.{RuleMatch, TextBlock, TextRange}
 
 object Text {
   def getSurroundingText(text: String, from: Int, to: Int): (String, String) = {
@@ -16,3 +18,4 @@ object Text {
   def getMatchTextSnippet(before: String, matchedText: String, after: String) =
     before + "[[" + matchedText + "]]" + after
 }
+
