+++ apps/rule-manager/app/utils/RuleManagerConfig.scala
@@ -9,6 +9,4 @@ class RuleManagerConfig(playConfig: Configuration, region: String, identity: App
   val dbUrl = playConfig.get[String]("db.default.url")
   val dbUsername = playConfig.get[String]("db.default.username")
   val dbPassword = playConfig.get[String]("db.default.password")
-  val credentials = playConfig.get[String]("typerighter.google.credentials")
-  val spreadsheetId = playConfig.get[String]("typerighter.sheetId")
 }
