+++ docs/01-running-locally.md
@@ -1,19 +1,8 @@
-# Dependencies
-We will need dev-nginx: `brew install dev-nginx`
-
-# Set up
-
-Run `dev-nginx setup-app ./nginx/nginx-mapping.yml` from the root of the project.
-
 # Basic developer run
 
-Get credentials for Composer profile.
+Get credentials for Composer profile then run `./script/start` and visit `localhost:9000` to confirm the service came up correctly.
 
-To launch the Checker, run `./script/start-checker` (`--debug` to run a debugger on port 5005).
-
-To start the Rule Manager, run `./script/start-manager` (`--debug` to attach a debugger on port 5006) and visit [the locally running app](https://manager.typerighter.local.dev-gutools.co.uk) to check it is running.
-
-To run everything, run `./script/start` (`--debug` will attach debuggers on both services in the respective ports above) and visit [the locally running app](https://manager.typerighter.local.dev-gutools.co.uk/) to confirm the service came up correctly.
+To attach an interactive debugger, run `./script/start --debug` to expose port 5005 for debugging.
 
 If you're testing changes to the rule audit client, see [the additional steps in its README](https://github.com/guardian/typerighter/tree/main/rule-audit-client).
 
@@ -60,7 +49,3 @@ EOF
 ```
 
 Assuming you are running a CODE environment, the response should indicate that the word 'Grauniad' is misspelled.
-
-## Running the rule manager
-
-`sbt ruleManager/run`
