--- script/lib/check-aws-credentials
@@ -1,12 +1,0 @@
-#!/bin/bash
-red='\033[0;31m'
-plain='\x1B[0m'
-
-STATUS=$(aws sts get-caller-identity --profile composer 2>&1 || true)
-if [[ ${STATUS} =~ (ExpiredToken) ]]; then
-echo -e "❌Credentials for the composer profile are expired. Please fetch new credentials and run this again.❌"
-exit 1
-elif [[ ${STATUS} =~ ("could not be found") ]]; then
-echo -e "❌Credentials for the composer profile are missing. Please ensure you have the right credentials.❌"
-exit 1
-fi
